feat: moving around layout and layout shell
