+++ src/app/layout.tsx
@@ -1,8 +1,6 @@
 import '../../styles/globals.scss';
 import type { Metadata } from 'next';
-import { Header } from '@/components/Header';
-import { Footer } from '@/components/Footer';
-import { PhaseBanner } from '@/components/PhaseBanner';
+import { LayoutShell } from '@/app/layoutShell';
 
 export const metadata: Metadata = {
   title: {
@@ -13,24 +11,6 @@ export const metadata: Metadata = {
     'The Ministry of Justice Developer Portal — documentation, products, and guidelines for cross-government developers.',
 };
 
-export function LayoutShell({ children }: { children: React.ReactNode }) {
-  return (
-    <>
-      <a href="#main-content" className="govuk-skip-link" data-module="govuk-skip-link">
-        Skip to main content
-      </a>
-      <Header />
-      <PhaseBanner />
-      <main className="govuk-main-wrapper" id="main-content" role="main">
-        {children}
-      </main>
-      <Footer />
-    </>
-  );
-}
-
-// an example of how to ignore elements from coverage
-/* v8 ignore next */
 export default function RootLayout({ children }: { children: React.ReactNode }) {
   return (
     <html lang="en" className="govuk-template">
