+++ vitest.config.ts
@@ -22,7 +22,7 @@ export default defineConfig({
       all: true,
 
       include: ['src/**/*.{ts,tsx}'],
-      exclude: ['tests/**', 'src/types/types.ts'],
+      exclude: ['tests/**', 'src/types/types.ts', 'src/app/layout.tsx'],
     },
   },
 });
